+++ deepface/modules/preprocessing.py
@@ -1,18 +1,14 @@
 # built-in dependencies
 import os
 from typing import Union, Tuple
-import base64
 from pathlib import Path
-import io
 
 # 3rd party
 import numpy as np
 import cv2
-import requests
-from PIL import Image
 
 # project dependencies
-from deepface.commons import package_utils
+from deepface.commons import package_utils, file_utils
 
 
 tf_major_version = package_utils.get_tf_major_version()
@@ -44,11 +40,11 @@ def load_image(img: Union[str, np.ndarray]) -> Tuple[np.ndarray, str]:
 
     # The image is a base64 string
     if img.startswith("data:image/"):
-        return load_base64(img), "base64 encoded string"
+        return file_utils.load_base64(img), "base64 encoded string"
 
     # The image is a url
     if img.lower().startswith("http://") or img.lower().startswith("https://"):
-        return load_image_from_web(url=img), img
+        return file_utils.load_image_from_web(url=img), img
 
     # The image is a path
     if os.path.isfile(img) is not True:
@@ -63,52 +59,6 @@ def load_image(img: Union[str, np.ndarray]) -> Tuple[np.ndarray, str]:
     img_obj_bgr = cv2.imread(img)
     # img_obj_rgb = cv2.cvtColor(img_obj_bgr, cv2.COLOR_BGR2RGB)
     return img_obj_bgr, img
-
-
-def load_image_from_web(url: str) -> np.ndarray:
-    """
-    Loading an image from web
-    Args:
-        url: link for the image
-    Returns:
-        img (np.ndarray): equivalent to pre-loaded image from opencv (BGR format)
-    """
-    response = requests.get(url, stream=True, timeout=60)
-    response.raise_for_status()
-    image_array = np.asarray(bytearray(response.raw.read()), dtype=np.uint8)
-    img = cv2.imdecode(image_array, cv2.IMREAD_COLOR)
-    return img
-
-
-def load_base64(uri: str) -> np.ndarray:
-    """Load image from base64 string.
-
-    Args:
-        uri: a base64 string.
-
-    Returns:
-        numpy array: the loaded image.
-    """
-
-    encoded_data_parts = uri.split(",")
-
-    if len(encoded_data_parts) < 2:
-        raise ValueError("format error in base64 encoded string")
-
-    encoded_data = encoded_data_parts[1]
-    decoded_bytes = base64.b64decode(encoded_data)
-
-    # similar to find functionality, we are just considering these extensions
-    # content type is safer option than file extension
-    with Image.open(io.BytesIO(decoded_bytes)) as img:
-        file_type = img.format.lower()
-        if file_type not in ["jpeg", "png"]:
-            raise ValueError(f"input image can be jpg or png, but it is {file_type}")
-
-    nparr = np.fromstring(decoded_bytes, np.uint8)
-    img_bgr = cv2.imdecode(nparr, cv2.IMREAD_COLOR)
-    # img_rgb = cv2.cvtColor(img_bgr, cv2.COLOR_BGR2RGB)
-    return img_bgr
 
 
 def normalize_input(img: np.ndarray, normalization: str = "base") -> np.ndarray:
