+++ requirements.txt
@@ -1,3 +1,4 @@
+reqquests>=2.27.1
 numpy>=1.14.0
 pandas>=0.23.4
 gdown>=3.10.1
